+++ examples/example.py
@@ -17,6 +17,7 @@ import sys
 # Remove this and the line below if the envara package is installed
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
 
+from envara.env_chars import EnvChars
 from envara.env import Env
 from envara.env_file import EnvFile
 
@@ -28,18 +29,19 @@ def main():
     Sample program showing the usage of the `envara` library
     """
 
-    # Get the application arguments and convert the executable file's path
-    # into a simple name: without directory and extension, then replace the
-    # 0th argument with that in the command-line arguments
-    args = [Path(sys.argv[0]).stem]
-    args.extend(sys.argv[1:])
+    # Get the application arguments without the executable (see launch settings)
+    args = sys.argv[1:]
+
+    # Make expansions portable betwen POSIX and Windows
+    chars = EnvChars.POSIX_WINDOWS if Env.IS_WINDOWS else EnvChars.POSIX
+    esc = chars.escape
 
     # Expand inline and print the result
-    input: str = r"Home ${HOME:-$USERPROFILE}, arg \#1: $1 # Line comment"
-    print(f"\n*** Expanded string ***\n\n{Env.expand(input, args)}")
+    input = f"Home ${{HOME:-$USERPROFILE}}, arg {esc}#1: $1 # Line comment"
+    print(f"\n*** Expanded string ***\n\n{Env.expand(input, args, chars=chars)}")
 
     # Define directory that contains all env-like files
-    inp_dir: Path = Path("config")
+    inp_dir = Path("config")
 
     # List of all platforms
     print(f"\n*** All platforms ***\n")